LGNAS Patch

+++ fs/udf/super.c
@@ -613,7 +613,9 @@ static loff_t udf_check_vsd(struct super_block *sb)
 	else
 		sectorsize = sb->s_blocksize;
 
-	sector += (sbi->s_session << sb->s_blocksize_bits);
+	// pch_110609 : integer type can cause overflow on dual-layer disc, 
+	//              so it's needed to convert long long type before shift operation 
+	sector += (((loff_t)sbi->s_session) << sb->s_blocksize_bits);
 
 	udf_debug("Starting at sector %u (%ld byte sectors)\n",
 		  (unsigned int)(sector >> sb->s_blocksize_bits),
